♻️ refactor: 폴더 구조 정리
