+++ src/app/(auth)/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '@/components/login/LoginForm';
+import LoginForm from '@/components/auth/LoginForm';
 import Image from 'next/image';
 import Link from 'next/link';
