+++ src/components/auth/Terms.tsx
